+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Cleanup httpx client usage.
   This is a breaking change to the HTTP client interface since we stopped passing the deprecated `data` argument to the httpx client methods (PUT & POST) and pass the `content` argument instead.
+- Add more type annotations to models and add a couple of `__str__` methods to models.
 
 ## 1.1.1 - 08-11-2023